dp inj

+++ Soccer IQ Demo/Controllers/ClubController.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Soccer_IQ_Demo.Models;
 using Soccer_IQ_Demo.Repository;
 
@@ -9,78 +8,56 @@ namespace Soccer_IQ_Demo.Controllers
     [ApiController]
     public class ClubController : ControllerBase
     {
+        private readonly ClubRepository clubRepository;
 
+        public ClubController(ClubRepository clubRepository)
+        {
+            this.clubRepository = clubRepository;
+        }
 
-        ClubRepository clubRepository = new ClubRepository();
+        [HttpGet]
         public IActionResult Display()
         {
-
-            var Clubs = clubRepository.GetAll();
-            return Ok(Clubs);
-
-
-
-
-
+            var clubs = clubRepository.GetAll();
+            return Ok(clubs);
         }
-        public IActionResult Create(Club club)
+
+        [HttpPost]
+        public IActionResult Create([FromBody] Club club)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                clubRepository.Add(club);
-                clubRepository.Commit();
-
-                return BadRequest();
+                return BadRequest(ModelState);
             }
-            return Ok(club);
 
+            clubRepository.Add(club);
+            return CreatedAtAction(nameof(Display), new { id = club.Id }, club);
         }
-        public IActionResult Edit(Club club)
+
+        [HttpPut]
+        public IActionResult Edit([FromBody] Club club)
         {
+            var existingClub = clubRepository.GetById(club.Id);
+            if (existingClub == null)
+            {
+                return NotFound("Club not found");
+            }
+
             clubRepository.Edit(club);
-            clubRepository.Commit();
-            return Ok();
+            return Ok(club);
         }
+
+        [HttpDelete("{clubId}")]
         public IActionResult Delete(int clubId)
         {
             var club = clubRepository.GetById(clubId);
+            if (club == null)
+            {
+                return NotFound("Club not found");
+            }
 
             clubRepository.Delete(club);
-            clubRepository.Commit();
-            return Ok();
-        }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+            return NoContent();
         }
     }
+}
